feat: ioc-v1.2.3.2 元对象的注册延后至QCoreApplocation构造时，而不是在静态变量初始话时. boot-v1.2.1.2 取消then的异步调用，改为同步. log-v1.0.0.3 优化代码和结构
