+++ McIocBoot/include/McBoot/McBootVersion.h
@@ -6,7 +6,7 @@
 #define MC_BOOT_MAJOR 1 // version when you make big feature changes.
 #define MC_BOOT_MINOR 2 // version when you make small feature changes.
 #define MC_BOOT_PATCH 1 // version when you make backwards-compatible bug fixes.
-#define MC_BOOT_INTERNAL 1 // MC_BOOT_INTERNAL version for perpurse like feature test, bug fix test,development, et
+#define MC_BOOT_INTERNAL 2 // MC_BOOT_INTERNAL version for perpurse like feature test, bug fix test,development, et
 
 #ifdef QT_DEBUG
 #define MC_BOOT_VERSION_STR (QString::number(MC_BOOT_MAJOR) + "." + QString::number(MC_BOOT_MINOR) + "." + QString::number(MC_BOOT_PATCH) + "." + QString::number(MC_BOOT_INTERNAL))
