+++ McIocBoot/src/Socket/McQmlSocket.cpp
@@ -98,6 +98,10 @@ void McQmlSocket::onMessage(const QJSValue &callback, bool isSync) noexcept
 
 void McQmlSocket::opened() noexcept 
 {
+    //! 2020-9-18
+    //! 由于js为单线程语言，只能在创建该对象的线程中调用回调函数
+    d->isOpenSync = true;
+    
     if(d->isOpenSync) {
         qApp->postEvent(this, new QmlSocketEvent(QmlSocketEvent::OpenEvent, ""));
     }else{
@@ -107,6 +111,10 @@ void McQmlSocket::opened() noexcept
 
 void McQmlSocket::closed() noexcept 
 {
+    //! 2020-9-18
+    //! 由于js为单线程语言，只能在创建该对象的线程中调用回调函数
+    d->isCloseSync = true;
+    
     if(d->isCloseSync) {
         qApp->postEvent(this, new QmlSocketEvent(QmlSocketEvent::CloseEvent, ""));
     }else{
@@ -116,6 +124,10 @@ void McQmlSocket::closed() noexcept
 
 void McQmlSocket::errored(const QString &errMsg) noexcept 
 {
+    //! 2020-9-18
+    //! 由于js为单线程语言，只能在创建该对象的线程中调用回调函数
+    d->isErrorSync = true;
+    
     if(d->isErrorSync) {
         qApp->postEvent(this, new QmlSocketEvent(QmlSocketEvent::ErrorEvent, errMsg));
     }else{
@@ -125,6 +137,10 @@ void McQmlSocket::errored(const QString &errMsg) noexcept
 
 void McQmlSocket::receivedMsg(const QVariant &msg) noexcept 
 {
+    //! 2020-9-18
+    //! 由于js为单线程语言，只能在创建该对象的线程中调用回调函数
+    d->isMessageSync = true;
+    
     if(d->isMessageSync) {
         qApp->postEvent(this, new QmlSocketEvent(QmlSocketEvent::MessageEvent, msg));
     }else{
