+++ McLogQt/include/McLog/Appender/impl/McFileDeviceAppender.h
@@ -1,18 +1,19 @@
 #pragma once
 
-#include "McAbstractAppender.h"
+#include "McAbstractFormatAppender.h"
 
-class MCLOGQT_EXPORT McFileDeviceAppender : public McAbstractAppender 
+class MCLOGQT_EXPORT McFileDeviceAppender : public McAbstractFormatAppender 
 {
     Q_OBJECT
     MC_DECL_INIT(McFileDeviceAppender)
-    MC_DEFINE_TYPELIST(MC_DECL_TYPELIST(McAbstractAppender))
+    MC_DEFINE_TYPELIST(MC_DECL_TYPELIST(McAbstractFormatAppender))
 public:
     Q_INVOKABLE McFileDeviceAppender();
     ~McFileDeviceAppender() override;
     
 protected:
-    void flush() noexcept override;
+    void writeBefore() noexcept override;
+    void writeAfter() noexcept override;
     
     virtual void tryNextFile() noexcept;
 };
